sketch of core refactor
