+++ src/zarr/core/array.py
@@ -63,6 +63,7 @@ from zarr.core.common import (
     product,
 )
 from zarr.core.config import config as zarr_config
+from zarr.core.group.async import AsyncGroup
 from zarr.core.indexing import (
     BasicIndexer,
     BasicSelection,
@@ -120,7 +121,7 @@ if TYPE_CHECKING:
 
     from zarr.abc.codec import CodecPipeline
     from zarr.codecs.sharding import ShardingCodecIndexLocation
-    from zarr.core.group import AsyncGroup
+    from zarr.core.group.async import AsyncGroup
     from zarr.storage import StoreLike
 
 
@@ -3730,7 +3731,7 @@ async def chunks_initialized(
 def _build_parents(
     node: AsyncArray[ArrayV2Metadata] | AsyncArray[ArrayV3Metadata] | AsyncGroup,
 ) -> list[AsyncGroup]:
-    from zarr.core.group import AsyncGroup, GroupMetadata
+    from zarr.core.group.metadata import GroupMetadata
 
     store = node.store_path.store
     path = node.store_path.path
