+++ tests/test_attributes.py
@@ -1,12 +1,13 @@
 import zarr.core
 import zarr.core.attributes
+import zarr.core.group.sync
 import zarr.storage
 
 
 def test_put() -> None:
     store = zarr.storage.MemoryStore()
     attrs = zarr.core.attributes.Attributes(
-        zarr.Group.from_store(store, attributes={"a": 1, "b": 2})
+        zarr.core.group.sync.Group.from_store(store, attributes={"a": 1, "b": 2})
     )
     attrs.put({"a": 3, "c": 4})
     expected = {"a": 3, "c": 4}
@@ -16,7 +17,7 @@ def test_put() -> None:
 def test_asdict() -> None:
     store = zarr.storage.MemoryStore()
     attrs = zarr.core.attributes.Attributes(
-        zarr.Group.from_store(store, attributes={"a": 1, "b": 2})
+        zarr.core.group.sync.Group.from_store(store, attributes={"a": 1, "b": 2})
     )
     result = attrs.asdict()
     assert result == {"a": 1, "b": 2}
