+++ tests/test_metadata/test_v3.py
@@ -11,7 +11,8 @@ from zarr.codecs.bytes import BytesCodec
 from zarr.core.buffer import default_buffer_prototype
 from zarr.core.chunk_key_encodings import DefaultChunkKeyEncoding, V2ChunkKeyEncoding
 from zarr.core.config import config
-from zarr.core.group import GroupMetadata, parse_node_type
+from zarr.core.group.core import parse_node_type
+from zarr.core.group.metadata import GroupMetadata
 from zarr.core.metadata.v3 import (
     ArrayV3Metadata,
     DataType,
